+++ instrumentation/opentelemetry-instrumentation-urllib/tests/test_urllib_integration.py
@@ -24,14 +24,14 @@ from urllib.request import OpenerDirector
 import httpretty
 
 import opentelemetry.instrumentation.urllib  # pylint: disable=no-name-in-module,import-error
-from opentelemetry import context, trace
-
-# FIXME: fix the importing of this private attribute when the location of the _SUPPRESS_HTTP_INSTRUMENTATION_KEY is defined.
-from opentelemetry.context import _SUPPRESS_HTTP_INSTRUMENTATION_KEY
+from opentelemetry import trace
 from opentelemetry.instrumentation.urllib import (  # pylint: disable=no-name-in-module,import-error
     URLLibInstrumentor,
 )
-from opentelemetry.instrumentation.utils import _SUPPRESS_INSTRUMENTATION_KEY
+from opentelemetry.instrumentation.utils import (
+    suppress_http_instrumentation,
+    suppress_instrumentation,
+)
 from opentelemetry.propagate import get_global_textmap, set_global_textmap
 from opentelemetry.sdk import resources
 from opentelemetry.semconv.trace import SpanAttributes
@@ -255,26 +255,16 @@ class RequestsIntegrationTestBase(abc.ABC):
         self.assert_span()
 
     def test_suppress_instrumentation(self):
-        token = context.attach(
-            context.set_value(_SUPPRESS_INSTRUMENTATION_KEY, True)
-        )
-        try:
+        with suppress_instrumentation():
             result = self.perform_request(self.URL)
             self.assertEqual(result.read(), b"Hello!")
-        finally:
-            context.detach(token)
 
         self.assert_span(num_spans=0)
 
     def test_suppress_http_instrumentation(self):
-        token = context.attach(
-            context.set_value(_SUPPRESS_HTTP_INSTRUMENTATION_KEY, True)
-        )
-        try:
+        with suppress_http_instrumentation():
             result = self.perform_request(self.URL)
             self.assertEqual(result.read(), b"Hello!")
-        finally:
-            context.detach(token)
 
         self.assert_span(num_spans=0)